# feat(precompiles): optimize bn128 precompiles