--- a/.gitignore
+++ b/.gitignore
@@ -57,3 +57,4 @@ Wallet
 
 /framework/propPath
 .cache
+bin