--- a/chainbase/build.gradle
+++ b/chainbase/build.gradle
@@ -37,7 +37,7 @@ dependencies {
     compile "org.fusesource.jansi:jansi:$jansiVersion"
     compile group: 'org.rocksdb', name: 'rocksdbjni', version: '5.15.10'
     compile group: 'com.typesafe', name: 'config', version: '1.3.2'
-    compile 'io.github.tronprotocol:zksnark-java-sdk:1.0.0'
+    compile 'com.github.zkbob:zksnark-java-sdk:feature~arkworks_alt_bn128-SNAPSHOT'
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.8.5'
     compile project(":protocol")
     compile project(":common")
@@ -86,6 +86,11 @@ jacocoTestReport {
             fileTree(dir: it,)
         }
     }
+}
+
+repositories {
+    // ...
+    maven { url "https://jitpack.io" }
 }
 
 build.dependsOn jacocoTestReport