… framework/src/test/resources/json/bn256Add.json
… framework/src/test/resources/json/bn256Pairing.json
… framework/src/test/resources/json/bn256ScalarMul.json