**What does this PR do?**
This PR optimizes `BN128Addition`, `BN128Multiplication`, and `BN128Pairing` precompiled contracts by using [arkworks](https://docs.rs/ark-bn254/latest/ark_bn254/) implementation of bn128 operations. This PR uses `LibarkworksWrapper` introduced in https://github.com/tronprotocol/zksnark-java-sdk/pull/9.

**Why are these changes required?**

The native implementation of these precompiles is relatively slow. In general, this hinders on-chain zkSNARKs verification and, consequently, makes zk-based apps almost unfeasible (https://github.com/tronprotocol/java-tron/issues/4311). With this PR (or a similar approach), these applications become viable. This PR is necessary to resolve this [issue](https://github.com/tronprotocol/java-tron/issues/5492).

It is important to note that these changes are not proposed to solve a specific problem of ZkBob but rather enable a whole class of applications and protocols based on elliptic curve cryptography and pairings in particular.  The possibilities brought by ZKP are not limited to privacy focused protocols ,  some other examples are zk based auth for AA wallets, light clients, zk based bridges, computational integrity enforcing contracts etc.

**This PR has been tested by:**
- Unit Tests (test cases were copied from the `go-ethereum`)
- Manual Testing

**Follow up**

**Extra details**
I've added some benchmarks before (https://github.com/zkBob/java-tron/tree/bn128-native-bench) and after modifications (https://github.com/zkBob/java-tron/tree/bn128-arkworks-bench). The results can be found below.

The average time of operations before (Intel(R) Core(TM) i7-10750H CPU, 32 GB RAM):
```
BN128Addition: 66387 ns
BN128Multiplication: 3553350 ns
BN128Pairing (10 pairs): 101565419 ns
```

The average time of operations after:
```
BN128Addition: 11576 ns
BN128Multiplication: 181301 ns
BN128Pairing (10 pairs): 3285601 ns
```