@r0wdy1 2023-09-21
@tomatoishealthy, thanks for description, these concerns seem reasonable. 
Apart from main net is it possible to make estimates about any public testnet deployments? 
Thank you.