@lvs007 2023-09-22
> > @tomatoishealthy, thanks for description, these concerns seem reasonable. Apart from main net is it possible to make estimates about any public testnet deployments? Thank you.
> 
> @r0wdy1 I agree with the idea.
> 
> 1. Confirm the final upgrade plan.
> 2. Private test network deployed and tested
> 3. Public test network deployed, fully tested and stabilized for a period of time
> 
> cc @tomatoishealthy @lvs007 @Federico2014 @ss3344520

Yes, a detailed code review is also required