@r0wdy1 2023-09-26
> In addition, no matter which solution we adopt, we need to retain the previous logic and use a way like if else to cooperate with optimization. The block height may need to be confirmed before upgrading and a new PR issued to update it

The purpose of this is to  avoid and necessity of simultaneous upgrade and let anyone to upgrade anytime before certain block height, right? 
Do you expect help from our team with this task?