@mumianhua 2023-09-26
the patch code coverage is lower than 80%, so the ci task is failed. 
please add some tests to cover more added code ! 
https://app.codecov.io/github/tronprotocol/java-tron/commit/118b0f183b362c66641a9904ecca96675c4f2017