@AllFi 2023-09-27
Hello, @mumianhua! I've added test cases with invalid data to cover most of the logic. The only code that is still uncovered is the handling of unsuccessful calls to `libarkworksAddG1` and `libarkworksMulG1`. Since there is preliminary validation of the input by calling `libarkworksG1IsValid` these methods should not fail with the current implementation. Perhaps, we can consider removing redundant checks but I think there is no harm in keeping them just in case.

The CentOS build has two failed tests but it seems that I didn't change anything related to them. Should I investigate this further, or are they possibly just flaky tests?