@lxcmyf 2023-10-23
@AllFi 
Based on this, I have the following concerns:

1. I noticed that `compile' com. github. zkbob: zksnark java sdk: feature~arkworks_ Alt_ Bn128 SNAPSHOT `  The static class LibarkworksJNI in the dependency uses **JNI** implementation, which may lead to potential risks such as memory management, buffer overflow, abnormal crash, code injection, etc. These risks have not been further verified, and it is not recommended to do so here
2. Some of the performance improvements you mentioned may be implemented in safer and more convenient ways, such as increasing timeout time, etc
3. The third-party dependency referenced by this PR has not been fully verified, and its security and stability are not yet clear
In summary, we have decided not to adopt the implementation of this PR temporarily. Thank you for your improvement suggestions and look forward to closer cooperation in the future.