@r0wdy1 2023-10-26
@lxcmyf @tomatoishealthy @halibobo1205 looking forward for your feedback and updated plan on the issue: we should either proceed with this PR or reopen https://github.com/tronprotocol/java-tron/issues/4311 and make some plan for timeout increase.
We're very excited about launching our DApp on Tron and this problem is the last roadblock in our way